:recycle: Split mod.ts roughly
